Client added comment count to post

+++ server/Codeus/src/main/java/orbitaljin/codeus/store/services/CommentService.java
@@ -2,6 +2,7 @@ package orbitaljin.codeus.store.services;
 
 import orbitaljin.codeus.store.models.Comment;
 import orbitaljin.codeus.store.repositories.CommentRepository;
+import orbitaljin.codeus.store.repositories.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +12,8 @@ import java.util.List;
 public class CommentService {
     @Autowired
     private CommentRepository repository;
+    @Autowired
+    private PostRepository postRepository;
 
     public List<Comment> getAllComments() {
         return repository.findAll();
@@ -18,18 +21,26 @@ public class CommentService {
 
     public Comment getCommentById(String id) {
         return repository.findById(id).orElse(null);
-
     }
 
     public Comment createComment(Comment comment) {
+        // Increment the comment count of the post
+        postRepository.findById(comment.getPostId()).ifPresent(post -> {
+            post.setCommentCount(post.getCommentCount() + 1);
+            postRepository.save(post);
+        });
         return repository.save(comment);
     }
 
-    public Comment updateComment(Comment comment) {
-        return repository.save(comment);
-    }
 
     public void deleteComment(String id) {
+        // Decrement the comment count of the post
+        Comment comment = repository.findById(id).orElse(null);
+        if (comment == null) return;
+        postRepository.findById(comment.getPostId()).ifPresent(post -> {
+            post.setCommentCount(post.getCommentCount() - 1);
+            postRepository.save(post);
+        });
         repository.deleteById(id);
     }
